Relocate module loader file
